perf: replace dlclark/regexp2 with regexp
The first change was to make the regular expressions declared within
the module `encoding.go` compatible with go's re2 engine by dropping
use of the idiom `(?!\S)`, which I am not really sure why we would
need it given where in most expressions it was being used.

And another noticeable change is how the index handling functions like
`findRegex2StringIndex` had to be rewritten.  There's quite a bit of
allocation there.  We may be able to squeeze even more perf outta
there.